GinkgoPython/pyproject.toml
[project]
name = "ginkgopython"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "accelerate>=1.12.0",
    "bitsandbytes>=0.49.2",
    "fastapi>=0.129.0",
    "language-data>=1.4.0",
    "optimum-quanto>=0.2.7",
    "pycountry>=26.2.16",
    "pydantic-settings>=2.0.0",
    "sqlmodel>=0.0.16",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "transformers>=5.2.0",
    "uvicorn[standard]>=0.41.0",
]

[dependency-groups]
dev = [
    "hf-xet>=1.2.0",
    "huggingface-hub>=1.4.1",
]

[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu130"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch" }
